Database Selections
We started coding the login form. We added verification and have started to select the database. In the regitserintern file we created a query to select the emails and check for duplicates. If the duplicate is found it is sent to the users browser. We also set up code to tell the user their registration has been made and it shows them their ID number.