Grabbing ID's
In the register page we added a hidden form into the page. In verify login we added a query to look for a user's username and password. If it was successful we placed the internID into a new variable along with the name. Then  we displayed a greeting to the user. After that we made a new page that finds the user and checks the ID.