Creating  a Table
I fixed some errors that were in my code. Then I started to work on the table on the SelectSeminar page and created another table for the seminar info.I have to put the data in the table. I added code to select the database and get the correct info into an array. I also need to change the code to the new table.